+++ Associations/book.rb
@@ -1,8 +1,7 @@
 require_relative '../Associations/rental'
 
 class Book
-  attr_reader :rentals
-  attr_accessor :title, :author
+  attr_accessor :title, :author, :rentals
 
   def initialize(title, author)
     @title = title
