+++ Entities/person.rb
@@ -5,8 +5,8 @@ require_relative '../Decorate/base_decorator'
 require_relative '../Associations/rental'
 
 class Person < Nameable
-  attr_reader :id, :rentals
-  attr_accessor :name, :age
+  attr_reader :id
+  attr_accessor :name, :age, :rentals
 
   def initialize(age, name = 'Unknown', parent_permission: true)
     super()
